requirements.txt
Flask==3.0.0
python-dotenv==1.0.0
openai==2.16.0
assemblyai==0.49.0
flask-sqlalchemy==3.1.1
Werkzeug==3.0.1
httpx==0.27.0
gunicorn==21.2.0
pypdf==4.0.1
ebooklib==0.18
python-docx==1.1.0
beautifulsoup4==4.12.3
youtube-transcript-api==1.2.4
